pyproject.toml
[project]
name = "poc-data-eng"
version = "0.1.0"
description = "POC: Renewable energy data platform — Airflow + Kafka + Databricks Delta Lake"
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "Proprietary" }
authors = [{ name = "Fuad Oñate", email = "fuadonates@gmail.com" }]

dependencies = [
    # Transformation
    "dbt-core>=1.9,<2.0",
    "dbt-databricks>=1.9,<2.0",
    # Orchestration
    "apache-airflow>=2.10,<3.0",
    # Streaming
    "confluent-kafka>=2.6,<3.0",
    # Data Quality
    "great-expectations>=1.3,<2.0",
    # Databricks
    "databricks-sql-connector>=4.0,<5.0",
    "databricks-sdk>=0.38,<1.0",
    # Data
    "polars>=1.20,<2.0",
    "pandas>=2.2,<3.0",
    "pyarrow>=18.0,<19.0",
    # HTTP / integrations
    "requests>=2.32,<3.0",
    "tenacity>=9.0,<10.0",
    # Utils
    "python-dotenv>=1.0,<2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8,<1.0",
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "pre-commit>=4.0,<5.0",
    "mypy>=1.13,<2.0",
    "sqlfluff>=3.2,<4.0",
    "responses>=0.25,<1.0",
    "types-requests>=2.32,<3.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "C4", "SIM", "PT"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers --tb=short"

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
ignore_missing_imports = true

[tool.sqlfluff.core]
dialect = "databricks"
templater = "jinja"
max_line_length = 120
